+++ packages/setupTests.ts
@@ -9,4 +9,6 @@ import '@testing-library/jest-dom/extend-expect';
 
 import { configure } from '@testing-library/react';
 
+import 'jest-canvas-mock';
+
 configure({ testIdAttribute: 'data-test-id' });