+++ packages/slider-with-histogram/src/components/Histogram.types.tsx
@@ -1,0 +1,16 @@
+interface Point {
+  x: number;
+  y: number;
+}
+
+interface HistogramProps {
+  values: number[];
+  width: number;
+  height: number;
+  selectedRangeIndexes?: [number, number];
+  normaliseData?: boolean;
+  showPoints?: boolean;
+  selectedRangeMinWidth?: number;
+}
+
+export { HistogramProps, Point };