+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# [9.0.0](https://github.com/hey-car/heycar-uikit/compare/v8.8.0...v9.0.0) (2023-06-14)
+
+
+### Bug Fixes
+
+* **header:** reset subnav state on page change + position fixed ([#219](https://github.com/hey-car/heycar-uikit/issues/219)) ([0430432](https://github.com/hey-car/heycar-uikit/commit/04304320dac9633f52e015aef6920889ab222816))
+
+
+### BREAKING CHANGES
+
+* **header:** Header changed from position sticky to position fixed
+
+HEYUI-365
+
+* update docs
+
+# [8.8.0](https://github.com/hey-car/heycar-uikit/compare/v8.7.0...v8.8.0) (2023-06-13)
+
+
+### Features
+
+* **sliderwithhistogram:** add slider with Histogram component  ([#218](https://github.com/hey-car/heycar-uikit/issues/218)) ([67be031](https://github.com/hey-car/heycar-uikit/commit/67be03198e4d660a56aabbbb64c7b1fbc23d463f))
+
 # [8.7.0](https://github.com/hey-car/heycar-uikit/compare/v8.6.1...v8.7.0) (2023-06-08)